Updates based on review feedback
- moving IAuthSession to AuthSession
- typing `props` as `IRequestProps` rather than casting with `as`
